Disable database backend
We need to prioritise a working build out of the box and the database doesn't
appear to build out of the box and when made to work it requires a bunch of
parameters when running the rts. It should be easy for users to run a
hello-world program and thus, we need to also make the database parameters
optional at run time. The default should be to not use the distributed database
backend. So, DB backend is now disabled. Let's re-enable when these outstanding
issues have been addressed!