Add basic test suite with one test
The README included instructions for running tests but there was no test
directory in the repo, so that never worked... or did someone forget to add
their local test stuff into the repo?

Anyhow, I ran into a funny bug trying to write my first acton program so I
added a basic test case for that.