Align space & exit when done
Changed some of the examples to use 4 space for indent consistently.

Also adding env.exit(0) at end so the program doesn't hang indefinitely. This
makes it possible to reuse the program in our test suite and since they aren't
meant to be used as long running daemons anyway, this is fine.