Use some examples in test suite
This is a bit of a hack to quickly expand the test suite; we reuse some of the
examples we have as tests of Acton. In the future, this could server to verify
that our examples are correct, rather than that the compiler is correct.
However, by that time we probably want to arrange this differently..

Only a subset of the examples are used since some rely on numpy, which
currently doesn't compile and I'm unsure about the suitability of some of the
others for use as tests. Starting with the low hanging fruit.