Ref actonc & don't hide errors
It used to be that actonc would output a bunch of debug message on
stderr and thus stderr was suppressed. That is no longer the case, so we
can remove the redirect to /dev/null... which in this case also shows
that we don't have actonc in our path. The install dir today is our
parent dir, so let's just use that for now. We should probably figure
out something better in the future.