Fix gcc linking
The order in which arguments are provided matters. gcc/ld will look for
things to link based on the compiled source files, thus the sources
should come before the linking args. Putting the libs to link first
becomes like a NOP since we have no unresolved references at that time.

I'm not sure why, guessing it has to do with different implementations
of gcc/ld, but this was working on OS X already before but trying to use
actonc under Linux would result in a lot of "undefined reference to X",
which are now all fixed.

Since we need -lkqueue on Linux, but there is no such library on OS X,
we use preprocessing to make compilation of actonc conditional based on
the host OS.