Revamp regression tests
I restructured how we run our regression tests. Now also testing that
the ones that are marked as failing are actually failing. I found that
one - string-concat, is working, despite previously being marked as
disabled.

We no longer run the examples as tests. I found it easier to structure
it like this.

I moved all the current tests to test/regression. All the tests in there
are expected to behave the same way, which is to compile with the actor
'main' as the root actor and they should all exit with return code 0 on
success. This makes it easy to write the Makefile.

We will have other more advanced tests, which we can now place in some
other subdirectory in test/ without colliding.

The name regression isn't entirely correct... but I couldn't come up
with anything better.

I think we should have a github issue that tracks each of these failing
cases!

Added a new test case for failing subtraction - Acton can't count :/

Also removing copyright from tests - seems a little heavy!? I don't
think it is needed.