rts: Add runtime args for DDB
It is now possible to configure the RTS to use the distributed database
backend for storing the run time state of actors. It was previously
enabled through a build time flag and the database host and port was
hard-coded. There are two run time options:

--rts-ddb-host, the hostname to connect to
--rts-ddb-port, the TCP port to connect to, default is 32000

A third option, --rts-verbose, was also added to make the RTS a bit more
chatty about it is doing, in particular in regards to the backend.

examples/count is probably the simplest program that illustrates
the persistence and restoration of state. In the below example, we first
start count, with the backend host set. It counts to 4 and we
stop it with ^C. On a second invocation, it will now first check for
existing actor state in the database, restore that and run the program,
thus we start counting from where we last left off:

    kll@Boxy:~/acton/examples$ ./count --rts-ddb-host localhost --rts-verbose
    Acton RTS: using distributed database backend (DDB): localhost:32000
    Acton RTS: checking for previous actor state in DDB... done
    0
    1
    2
    3
    4
    ^C
    kll@Boxy:~/acton/examples$ ./count --rts-ddb-host localhost --rts-verbose
    Acton RTS: using distributed database backend (DDB): localhost:32000
    Acton RTS: checking for previous actor state in DDB... done
    Acton RTS: restoring actor state from DDB... done
    5
    6
    7
    8
    ^C
    kll@Boxy:~/acton/examples$

This presumes that the database is running, which can be started
with (for a quorum):

    ./server -p 32000 -m 34000 -s 127.0.0.1:34000
    ./server -p 32001 -m 34001 -s 127.0.0.1:34000
    ./server -p 32002 -m 34002 -s 127.0.0.1:34000

The WITH_BACKEND build flag has been removed since it's not needed
anymore.

This is so f***** cool!