Add simple time module
This adds a time module to Acton. It exposes a single function, time(x),
which returns the current time as an integer with nanosecond precision.

time(x) currently takes an argument, though its value is irrelevant.
It's a dummy. I tried to remove the argument but  that breaks the build
and I don't get why. Halp.

I'm not sure how we should design our time module. Like we can model it
directly on Python's time module or we can do something else...
something better!? It's low level enough that I'm not sure it matters
making it more Pythonic / Actonic.

There is a simple test of time.time(x).