Add test cases for arithmetic operations
There is a bug in integer subtraction and this adds a test case to prove
it and others for arithmetics overall.

It is split up like this because I could make subtraction work by
including a test for division in the same file. How can that be?

A previous regression test for the broken subtraction has been removed,
since these tests replace it and are better structured.