Add release workflow
The test workflow, when run on the main branch, produces a release which
has been known as `latest`. It is now called `tip` to avoid confusion
when talking about the "latest release" - would that be `latest` or the
latest release with a proper version number? `tip` is easier and still
distinguishes this from a nightly, which it is not.

The new release workflow is essentially a copy of the test action but with
smaller variations like the tag is not statically set to 'tip' (was
'latest') but is based on the tag that is pushed. The trigger is of
course the pushing of a tag rather than any branch.

The test action is changed to ignore tags. The pre-release-tip job is
also changed to once again (I used this in early testing) use eine/tip,
since that action? will force push the tip tag. We still need the other
release creation job to set the title, body etc though. It's a little
redundant but easier than writing our own thing that fixes it all.