Include date & time in tip release filename
tip releases already include the date and time in the version, like so:

  $ ./actonc --version FOO
  acton 0.4.1.20210805.13.46.55

The filename of a tip release would not reflect that, it would just be
called like `acton-linux-x86_64-0.4.1.tar.bz2`. This changes so that the
filename will be `acton-linux-x86_64-0.4.1.20210805.13.46.55.tar.bz2`,
which should make it easier for folks that often download tip releases
to differentiate them from each other.