Revamp argument parsing
This addresses two concerns with the original, somewhat naive, argument
parsing, namely that:
- getopt_long() will reorder the arguments in argv
- we would pass all arguments to the Acton program

We now tell getopt_long() to return arguments in order. This likely
isn't super important for any Acton program anyway, but we might as well
do it. It's done by starting optstring with "-".

Second, we now consume the command line arguments that are meant for the
RTS and will not pass them on the to the Acton program. For example, if
we run:

  ./app foo --rts-verbose --bar --rts-verbose

in the Acton application env.argv would be: [./app, foo, --bar]

It's rather unlikely that there would be actual conflicts, but let's
just say someone actually wanted to pass in a literal '--rts-verbose' to
their Acton application, that would be possible by doing:

  ./app foo --rts-verbose --bar -- --rts-verbose

In which case the application env.argv would be:
[./app, foo, --bar, --, --rts-verbose]

A small test case is added to show argument parsing works.

We don't free any malloced resources since we need them pretty much
throughout the program. Could there be a situation where it makes sense
to free() or just for cleanliness?