Separate build output from sources
This splits up the source of Acton from the built output, which goes in
the dist directory. Baking a release is then a simple operation of
taring up the dist directory.

When we invoke actonc, like for building tests, we do that in the dist
directory, thus producing output in that directory rather than mixed up
with the source code of Acton itself.

The release will be slightly slimmer, skipping certain .o files, .c
files etc that were previously included.