Add acton.rts.sleep() function
Which implies we are adding the acton.rts module too! There's a simple
test case added.

I don't know where to write documentation, like it would be nice to have
a docstring under the function but there's no function defined in Acton!
Just a function type signature, so no place to put the docstring from
which we can later generate documentation. I added a big comment for
now.

As stated in this comment, acton.rts.sleep() is sort of bad and
shouldn't normally be used, but it is deliberately implemented the way
it is and can be useful for benchmarking scenarios and similar.