Reorganize modules
This moves two of the standard library modules, math and time, into the
modules directory. They have previously been split out as the modules
directory acted both as a source and output location of actonc
compilation. That changed when we introduced the dist directory, to
which we write output - it then acts as the working directory of actonc.

The modules directory in the source git repository is now used for
source files and is not a working directory for actonc. Thus, we move
our standard library modules in there.

numpy is currently an exception as it consists of many files and is
inevitably much more compliated to compile, thus it remains as a
separate directory. I suspect most modules will follow the simpler
pattern of time and math, where .act, .h and .c file all have the same
prefix. Let's wait and see, if this turns into crap will come up with
some other structure.