Add random module
This adds a random module with a single function called randint(min,
max) that returns a pseudo-random number between min and max. It's using
the basic srand() / rand() for simplicity, portability and speed (we
don't need true randomness). To make it more idiomatic Pythonic I
decided to just use time() as the initial seed, which is set up when
initializing the module.

The code avoids skew and should return good random numbers. Not
cryptographically secure in any way but still not skewed in anyway.

The Acton RTS uses threads to execute actors. rand() is not thread safe,
but I don't think that matters here. AFAIK, since rand() is a PRNG, it
actually means rand() is a deterministic function based on the initial
seed. It is not thread safe and by that, I think it is meant that it
will not act deterministically when invoked across multiple threads,
since it shares state. However, given that the Acton RTS does not give
any guarantees on how actors are scheduled against RTS threads, and thus
OS threads, even a thread safe random library wouldn't be deterministic
at the actor level. Thus, this sort of thread safety does not matter.
The only important thing is that it doesn't outright crash when invoked
by two concurrent threads but I don't think that's the case, heh.

There's a simple test case that asks for 100k random values to ensure
they are not outside of the specified range. We also do a check to
ensure that we get a roughly even spread of numbers by asking for 1
million numbers between 500 and 1000. The result should be within 10% of
an even spread to pass. Certainly possible to make something much
fancier than that, but this is good enough for now I think.