Fix the example in README
We were previously referring to the examples/ directory, but that's not
included in a release tar ball. Now we instruct the user to write their
own helloworld.act file, which shows a little bit more. Not much - it is
still very simple - but enough to understand a bit more of the basics!