Fix database backed RTS bootstrap
The RTS when used with the distributed database backend reads in any
previously existing state from the database. If there is no state, it
should bootstrap. A bug crept in during the introduction of DDB as a run
time arg that made the RTS always bootstrap actors, regardless of any
previously found actor state in the database. This is now fixed.