Workaround build problem when running DDB
Linux (and likely other) systems will prevent copying a file to a
destination that is currently being run. I've run into this a bunch of
times with the actondb executable where I am running it locally for
testing and then run make in the acton repo, which copies backend/server
to dist/actondb. If actondb was running, the system would refuse this as
the system does not allow modifying a running program. By moving it in
place we work around the issue.