Reduce output clutter from running tests
Most of the test output, in particular when running the negative tests
that we expect to fail compilation, produce rather verbose output and it
can be difficult to differentiate an actual error from a currently
expected error (in a negative test). Thus, we now reduce this output
clutter by piping stdout and stderr to /dev/null. There's also a small
header in front of each test that makes it easier to keep them apart.