Add arg to treat warnings as errors
There are currently some warnings being produced when compiling the C
code generated by actonc. We believe this likely indicates bugs and
thus, by treating these warnings as errors we are able to detect such
situations earlier. However, we also think that we should detect these
warnings in CI or similar environments and not turn our user base into a
warning detection system (the user who upgrades to the latest gcc first
will get bitten!), thus we offer this as an opt-in argument called
--pedantic. It is enabled for some of our own tests in CI.
Unfortunately, our regression tests emit warnings which would make them
fail with --cpedantic and it is thus currently not used.