Do not clean away modules/math.h
Previously the math module lived in a math/ subdirectory and
modules/math.h was just a copy of math/math.h. We recently moved simple
modules like math to live directly in modules/, and thus, modules/math.h
is now the real file. We should not remove it when cleaning!