Pythonify time module
This changes the time.time() function to return a float, just as the
Python time module does. However, we still return nano second precision
instead of microsecond precision as in Python. In addition, we add a
time.time_ns() function that returns the current time as nanoseconds,
just like in Python.

I've also added the equivalent to Pythons monotonic time functions,
monotonic() and monotonic_ns().

We no longer do a hard exit but return an exception in case we can't get
the time.

Tests have been revamped somewhat.