Disabled test '105-type-inference-order'.
File 105-type-inference-order.act captured a type inference bug that has now been fixed. But the fix has now made the example code
reveal an independent error in the rts that is intermittent -- that is, dependent on worker thread scheduling by the os. I've found
the cause and will of course commit a fix, but this will have to be from another branch. Disabling the test that intermittently
fails in the meantime.