Add RTS debug mode
This enables various debug printing while the RTS is running. It is very
detailed and will most definitely have a major performance impact. Thus,
it can only be used for debugging purposes in controlled environments.
Debug output is enabled with the --rts-debug option.

To completely eliminate any overhead in normal operations the debug
prints are not just conditioned on the runtime --rts-debug flag but are
also conditioned behind a build time flag. We build the rts lib twice,
once with the flag enabled and once without. actonc chooses which one to
link with when compiling the .act file. That is, given the program
helloworld.act, to run with debug output, we first have to compile with:

  $ actonc --rts-debug --root main helloworld.act

Then also run the program itself with:

  $ ./helloworld --rts-debug

If helloworld was not compiled with --rts-debug, we will get an error:

  $ ./helloworld --rts-debug
  ERROR: RTS debug not supported.
  HINT: Recompile this program using: actonc --rts-debug ...
  $

The verbose messages we printed before have been converted to use a
rtsv_printf function that is similar to rtsd_printf but for verbose
information. It is not conditioned on the use of actonc --rts-debug but
is always available and meant for information that isn't printed in hot
loops where it will impact performance.

A few messages have been improved in formatting and content.

Further, error messages are now written to STDERR.