Support multiple database servers
Instead of just connecting to a single database server, we can now
connect to multiple servers! The port used is specified with
--rts-ddb-port but can be specified per server by supplying the port
colon separated, like --rts-ddb-host localhost:1234 for port 1234.

This doesn't support IPv6 addresses today, like [::1]:1234, but the
database backend only listens to AF_INET anyway, so we can fix that a
little later.