Rewrite builtin targets to use wildcards
Rather than enumerate all files, we can find them by wildcards. I'm not
sure if we need to add things here or not, but this makes it more robust
towards addition, although at the same time it is more fragile from junk
files or whatnot. Anyway, I think this is easier to maintain.