Update changelog for 0.5.3
We released v0.5.3 from a maintenance branch, so the CHANGELOG on main
drifted out of sync. This is bringing it up to par again.

Further, the version number is now bumped to 0.5.3 but note how we won't
build the 0.5.3 from the main branch as the release build is based on a
git tag.