Build on debian sid
Instead of running on pesky ubuntu, we switch to Debian :)

Debian sid is uh, well, bleeding edge. We probably want something a
little more stable than that preferably, but for now this is the easiest
way of getting a recent enough version of libkqueue-dev that we need.

Once we move to epoll or io_uring on linux, we can drop libkqueue-dev
and potentially move to another debian version.

Since we want to build .debs on Debian, I think it is good to align as
many jobs as possible on Debian for now. Keeps it simple.