Try milliseconds for fun and profit
It's weird. Using microseconds, as two commits ago, I saw the timer
event trigger all the time, like every 10 microseconds or so. Switching
to nanoseconds fixes this on my computer and I'm only getting triggered
when I'm supposed to. The test cases pass etc. However it fails in CI.
For sake of completeness, I'm now trying the default of passing
milliseconds. Let's see what CI thinks of it.