Remove dependency on libkqueue
The Acton RTS was written for kqueue on MacOS. We got it running on
Linux by using a shim called libkqueue which pretends to be kqueue
towards the application but translates these calls into epoll on Linux.

We ran into some problems, which we thought were related to libkqueue
but ultimately turned out to be something else. As we did ascribe those
problems to libkqueue, we resorted to replacing it with epoll.

It turned out that libkqueue wasn't the problem, but switching to native
epoll is a good thing, so all's well!