Release v0.6.0
Update changelog and bump version number.

I've changed how all the links work in the CHANGELOG.md. This seems to
be the recommended style and makes things look better when rendered as
the links are on the same line as the actual item instead of the next
line... We could of course fix this by not breaking lines in the
markdown, but I want to keep it readable as a raw text file as well, and
thus I want to keep line breaks.