Use a minimum of 4 worker threads
On very low CPU count machine, like with a single CPU, Acton would not
work at all as the first thread runs the eventloop and without a second
thread there will be no workers at all. We now force a minimum of 4
worker threads. CPU affinity is normally set on the threads but when we
do not have a 1:1 mapping between worker threads and we don't set
affinity at all and let the threads roam - up to the normal scheduler to
figure out!