Reorder workflow to avoid racy delete
We've seen that the recreation of the tip release can fail with an error
message about failing to update. The GitHub Action we use is capable of
creating or updating an existing release and as we deleted the release
just before, GitHubs eventually consistent API probably shows that it
exists for a brief glimpse of time afterwards.

By moving up the deletion to the first thing we do, and having some
things that actually take some time, like checking out the repository
and downloading artifacts, we will likely get rid of this race.