Move defs to numpy.h
The files we keep in stdlib/src will eventually end up in the output
distribution system path under types/. A .h/.ty file in there is
effectively a module in Acton (and we expect the object file in
libActon.a). This means that the code for a module should either be
contained to that single file (well, .c & .h & .act) or that we somehow
include more files in some other location.

In this case stdlib/src/numpy.h previously included numpy/init.h.
Without qualification, init.h would look like a module called init.
Clearly we can't have that. The numpy internal files need to be hidden
away and preferably, we want to package it up in such a way that what we
do at development time with source file doesn't need to bleed through
to the operational environment (the Acton system path).

We now achieve that by:
- placing the entire interface of the numpy module in stdlib/src/numpy.h
  instead of including 6 other header files
  - including those header files meant that we had to propagate them
    into the final installed Acton system and... it's just messy

We still leak through some of the module internals of numpy as we have
to link the final Acton program with libbsd on Linux. Preferably we
should partial link numpy.o with libbsd or perhaps include libbsd.o in
libActon.a. Statically linking in such libraries reduces run time
dependencies and makes Acton more deterministic, which is what we
want... but that's for later.