Improve actonc (parallel) build & fix deb version
DAG DAG DAG DAG! The compiler Makefile is now folded into the main
Makefile!

Turned the actonc make target into a non-phony target. Nice! I just hope
I've caught all the files that affect the stack build. Since various
other files, like .ty, depend on actonc, making it non-phony is good
since any target that depends on a phony is always run. Now make will
actually do nothing when there is nothing to do!

Also enabled parallel build of actonc itself using 4 threads. This
brings down the build time on my AMD 5950X from ~55 seconds (single
thread) to ~35 seconds. Higher values do not yield a noticeable
improvement. Given the completeness of the DAG, doing parallel make of
the whole project is quite fast. GitHub Actions don't have that many CPU
cores but 4 is low enough of a value to not trash everything with
context switching... I think.

The manner in which we generate the version number is revamped. We
previously generated it when compiling actonc and then let the rest of
the build query actonc --version to get the version number. In
particular when building a release, that version from actonc would get
inserted into the tar ball name. Debian packages didn't use the same
value for the version but rather just grabbed the VERSION variable from
common.mk, which works out okay for release builds but for tip builds it
wouldn't reflect the build time - bad. The tricky bit was that in a
debian build, the deb target is the first target that gets called, which
in turn sets up a new "clean room" build, so it takes the version number
as input. Since actonc isn't compiled yet, we can't query it!

We have now turned everything around, so we generate the version number
in the makefile. If we are just building actonc etc as normal, then
we'll insert that version number. If we are building a debian package,
then we'll take the version number and feed into the deb build. This
will in turn run the normal build, at which point we've feed in the
version number which will then be used.