Store all .deb related artifacts in deb dir
We used to only store the actual .deb file as an artifact but there are
more files produced as part of the .deb build target. We now store all
of them in a deb/ directory and save that directory as an artifact. This
should make it easier to restore and finally add the files to a apt repo.