Run CI for all PRS
"push" triggers for any push to the Acton repository. With the
triangular git flow, PRs will be opened from a branch on a fork. To run
CI for those PRs we cannot use the "push" trigger but will have to use
the pull_request trigger.

We now use both triggers, which means we will see "duplicate" tests run,
but that's okay for now.