Introduce dev compiler profile
We effectively do this by renaming the --rts-debug flag since we want to
condition more things on the more generic "dev mode".

There are two "compiler profiles";
- development
- release

There is currently no explicit option for release mode; it is implied by
not specifying development mode with --dev. Not sure if we need an
explicit flag to it. Keeping it simple for now.

The -g flag, i.e. inclusion of debug symbols, is now conditioned on the
--dev profile.