Add worker stats & RTS monitor
This adds statistics per RTS worker thread and the ability to expose
these statistics over a UNIX domain socket. It is enabled with
--rts-mon PATH to listen on the socket PATH. The protocol use is a
simple ASCII line based protocol. A client can send "WTS" (currently
only supported command) to which the RTS will respond with a dump of the
worker thread statistics (WTS) as a JSON blob on one line.

There's a new thread to deal with the monitor socket. We only accept a
single connection at a time. I want the monitoring thread to be
completely standalone, i.e. I do not want to mix in listening on various
fds into our normal eventloop. This is deliberately kept simple.

To encode data in JSON we use the yyjson library, which is a wicked fast
JSON encoding and decoding library. It only consists of a single .h and
.c file, so they are now included in the deps directory. Not sure if
this is how we want to deal with dependencies in general, but it is
there for now.

The statistics are:
- program name (argv[0])
- PID
- current worker state
  - one of no-exist, worker, idle or sleeping
- number of times a thread has gone to sleep
- number of executed continuations
- sum of time spent executing continuations, in nanoseconds
- buckets with execution time & "bookkeeping" of continuations
  - for example, if a continuation takes less than 100ns to run, it is
  counted in the 100ns bucket, if it is between 100ns and 1us it is
  counted in the 1us bucket and so forth
  - bookeeping includes flushing outgoing queues with the locking etc
    involved, if the distributed database is used, then its interaction
    is included
  - the buckets are:
    - < 100ns
    - < 1us
    - < 10us
    - < 100us
    - < 1ms
    - < 10ms
    - < 100ms
    - < 1s
    - < 10s
    - < 100s
    - < +Inf
  - 100ns is more on the level of measurement overhead so a 10ns or 1ns
    bucket would not yield useful information

It should be possible to graph the ratio between our bookkeeping and the
actual execution of threads to show the sort of run time overhead of the
Acton system. Going to be particularly interesting to see for the
distributed database!

There is a new utility in utils/actonmon which can connect to the
monitoring socket and display thread stats. It supports three different
modes:
- simple
- rich
- prometheus

The rich interface (started with --rich) uses the rich library to render
a table with the worker thread statistics. It updates in place and looks
pretty good. The update interval can be set with --interval, in seconds.

The simple mode just prints some statistics to the screen in a long log
at every --interval. It doesn't have any dependencies and so is useful
when you don't have rich installed. It's currently very sparse.

prometheus (--prom) mode starts to listen on http://localhost:8000 and
will answer GET queries with the statistics that it collects from the
RTS.

To start examples/count with RTS monitoring enabled:

    examples/count --rts-mon ~/act_mon_socket

Then check the mon utility with the rich interface:

    utils/actonmon ~/act_mon_socket --interval 0.2 --rich

Or to enable prometheus export:

    utils/actonmon ~/act_mon_socket --prom